requirements.txt
xml.etree.ElementTree
pathlib
tensorflow = 2.8.0
cv2 = 4.5.5.62
numpy
matplotlib.pyplot
os
sklearn = 1.0.2